Move NeovimSurface out to separate file
